move helper functions to service file
