+++ packages/front-end/components/CustomFields/constants.ts
@@ -5,11 +5,3 @@ export const CUSTOM_FIELD_SECTION_LABELS: Record<CustomFieldSection, string> = {
   feature: "Features",
   experiment: "Experiments",
 };
-
-export function isCustomFieldBooleanTrue(value: unknown): boolean {
-  return value === true || value === "true";
-}
-
-export function toCustomFieldBooleanString(value: boolean): string {
-  return value ? "true" : "false";
-}
